// src/auth/enums/permissions.enum.ts
export enum PermissionsEnum {
  ALL = 'ALL',
  CREATE_MANAGER = 'CREATE_MANAGER',
  UPDATE_MANAGER = 'UPDATE_MANAGER',
  FIND_MANAGER = 'FIND_MANAGER',
  EXPORT_MANAGER = 'EXPORT_MANAGER',
  CREATE_EXPERT = 'CREATE_EXPERT',
  UPDATE_EXPERT = 'UPDATE_EXPERT',
  FIND_EXPERT = 'FIND_EXPERT',
  EXPORT_EXPERT = 'EXPORT_EXPERT',
  CREATE_TEACHER = 'CREATE_TEACHER',
  UPDATE_TEACHER = 'UPDATE_TEACHER',
  FIND_TEACHER = 'FIND_TEACHER',
  EXPORT_TEACHER = 'EXPORT_TEACHER',
  UPLOAD_ADMISSIONS = 'UPLOAD_ADMISSIONS',
  CREATE_FIELD_OF_STUDY = 'CREATE_FIELD_OF_STUDY',
  UPDATE_FIELD_OF_STUDY = 'UPDATE_FIELD_OF_STUDY',
  FIND_FIELD_OF_STUDY = 'FIND_FIELD_OF_STUDY',
  EXPORT_FIELD_OF_STUDY = 'EXPORT_FIELD_OF_STUDY',
  FIND_PERMISSIONS = 'FIND_PERMISSIONS',
  ASSIGN_PERMISSIONS = 'ASSIGN_PERMISSIONS',
  EXPORT_STUDENT = 'EXPORT_STUDENT',
  FIND_STUDENT = 'FIND_STUDENT',
  CREATE_STUDENT = 'CREATE_STUDENT',
  UPDATE_STUDENT = 'UPDATE_STUDENT',
  GET_BILL = 'GET_BILL',
  GET_CV = 'GET_CV',
}